Rework
I've reworked it a bunch. Honestly the background loop worked but it
became really horrible and the locking behaviour a little sketchy. While
this will slow down some requests a bit, unless it becomes measurably
slow let's keep it that way rather than introducing a bit of a horrible
pattern.